Reorganize modules
Move low-level interface under `ISL.Native`, making the room for a new
high-level interface in `ISL` and `ISL.Types`.
