+++ src/ISL.hs
@@ -1,0 +1,4 @@
+-- | High-level interface to isl.
+module ISL where
+
+import ISL.Types
