+++ src/ISL/Types.hs
@@ -1,0 +1,2 @@
+-- | Types for the high-level interface to isl.
+module ISL.Types where
